+++ pymilo/utils/test_pymilo.py
@@ -13,6 +13,7 @@ from ..pymilo_func import compare_model_outputs
 from ..chains.linear_model_chain import is_linear_model
 from ..chains.neural_network_chain import is_neural_network
 from ..chains.decision_tree_chain import is_decision_tree
+from ..chains.clustering_chain import is_clusterer
 
 from ..pymilo_param import EXPORTED_MODELS_PATH
 
@@ -32,12 +33,14 @@ def pymilo_export_path(model):
         model_type = "NEURAL_NETWORK"
     elif is_decision_tree(model):
         model_type = "DECISION_TREE"
+    elif is_clusterer(model):
+        model_type = "CLUSTERING"
     else:
         model_type = None
     return EXPORTED_MODELS_PATH[model_type]
 
 
-def pymilo_test(model, model_name, test_data):
+def pymilo_test(model, model_name):
     """
     Return the pymilo imported model's outputs for given test_data.
 
@@ -45,12 +48,8 @@ def pymilo_test(model, model_name, test_data):
     :type model: any sklearn's model class
     :param model_name: model name
     :type model_name: str
-    :param test_data: data for testing
-    :type test_data: np.ndarray or list
     :return: imported model's output
     """
-    x_test, _ = test_data
-
     export_model_path = pymilo_export_path(model)
     exported_model = Export(model)
     exported_model_serialized_path = os.path.join(
@@ -59,7 +58,7 @@ def pymilo_test(model, model_name, test_data):
 
     imported_model = Import(exported_model_serialized_path)
     imported_sklearn_model = imported_model.to_model()
-    return imported_sklearn_model.predict(x_test)
+    return imported_sklearn_model
 
 
 def pymilo_regression_test(regressor, model_name, test_data):
@@ -80,7 +79,7 @@ def pymilo_regression_test(regressor, model_name, test_data):
         "mean-error": mean_squared_error(y_test, pre_pymilo_model_y_pred),
         "r2-score": r2_score(y_test, pre_pymilo_model_y_pred)
     }
-    post_pymilo_model_y_pred = pymilo_test(regressor, model_name, test_data)
+    post_pymilo_model_y_pred = pymilo_test(regressor, model_name).predict(x_test)
     post_pymilo_model_prediction_outputs = {
         "mean-error": mean_squared_error(y_test, post_pymilo_model_y_pred),
         "r2-score": r2_score(y_test, post_pymilo_model_y_pred)
@@ -110,7 +109,7 @@ def pymilo_classification_test(classifier, model_name, test_data):
         "accuracy-score": accuracy_score(y_test, pre_pymilo_model_y_pred),
         "hinge-loss": hinge_loss(y_test, pre_pymilo_model_y_pred)
     }
-    post_pymilo_model_y_pred = pymilo_test(classifier, model_name, test_data)
+    post_pymilo_model_y_pred = pymilo_test(classifier, model_name).predict(x_test)
     post_pymilo_model_prediction_outputs = {
         "accuracy-score": accuracy_score(y_test, post_pymilo_model_y_pred),
         "hinge-loss": hinge_loss(y_test, post_pymilo_model_y_pred)
@@ -121,7 +120,33 @@ def pymilo_classification_test(classifier, model_name, test_data):
     report_status(comparison_result, model_name)
     return comparison_result
 
+def pymilo_clustering_test(clusterer, model_name, x_test, support_prediction = False):
+    """
+    Test the package's main structure in clustering task.
 
+    :param clusterer: the given clusterer model
+    :type clusterer: any valid sklearn's clusterer class
+    :param model_name: model name
+    :type model_name: str
+    :param x_test: data for testing
+    :type x_test: np.ndarray or list
+    :param support_prediction: whether the given clusterer supports .predict function or 
+    :type support_prediction: boolean
+    :return: True if the test succeed
+    """
+    pre_pymilo_model = clusterer
+    post_pymilo_model = pymilo_test(clusterer, model_name)
+    if(support_prediction):
+        pre_pymilo_model_y_pred = pre_pymilo_model.predict(x_test)
+        post_pymilo_model_y_pred = post_pymilo_model.predict(x_test)
+        mse = ((post_pymilo_model_y_pred - pre_pymilo_model_y_pred)**2).mean(axis=0)
+        epsilon_error = 10**(-8)
+        return report_status(mse < epsilon_error, model_name) 
+    else:
+        # TODO, apply peer to peer 
+        # Evaluation: peer to peer field type & value check 
+        return report_status(True, model_name)
+    
 def report_status(result, model_name):
     """
     Print status for each model.